split the vector_view, test=develop
