+++ lite/model_parser/flatbuffers/CMakeLists.txt
@@ -9,6 +9,6 @@ lite_fbs_library(fbs_op_desc SRCS op_desc.cc FBS_DEPS framework_fbs_header)
 lite_fbs_library(fbs_var_desc SRCS var_desc.cc FBS_DEPS framework_fbs_header)
 lite_fbs_library(fbs_block_desc SRCS block_desc.cc FBS_DEPS framework_fbs_header)
 lite_fbs_library(fbs_program_desc SRCS program_desc.cc FBS_DEPS framework_fbs_header)
-lite_fbs_library(vector_view SRCS vector_view.cc FBS_DEPS framework_fbs_header)
 
-lite_cc_test(test_vector_view SRCS vector_view_test.cc DEPS vector_view)
+lite_cc_test(test_vector_view SRCS vector_view_test.cc)
+add_dependencies(test_vector_view framework_fbs_header)
